adds backend services to actions, only tasks related

+++ src/client/src/redux/actions/tasks_actions.js
@@ -1,13 +1,31 @@
 import actionTypes from "../constans";
+import ItemClient from "./../../services/taskService";
+
+const itemService = new ItemClient();
+
+export const getTasksAction = () => {
+  return async (dispatch) => {
+    const tasks = await itemService.fetchTasks();
+    const sortedTasks = tasks.sort((a, b) => a.position - b.position);
+    dispatch(setTasks(sortedTasks));
+  };
+};
 
 const addTask = (task) => ({
   type: actionTypes.ADD_TASK,
   payload: task,
 });
 
-export const addTaskAction = (task) => {
+export const addTaskAction = (input, position) => {
   return async (dispatch) => {
-    dispatch(addTask(task));
+    const addedTasks = await itemService.addTask(input, false, position);
+    if (addedTasks.status === 200) {
+      for (let i = 0; i < addedTasks.response.length; i++) {
+        dispatch(addTask(addedTasks.response[i]));
+      }
+    } else if (addedTasks.status === 400) {
+      //TODO: Need to post the toast message
+    }
   };
 };
 
@@ -17,8 +35,13 @@ const removeTask = (taskID) => ({
 });
 
 export const removeTaskAction = (id) => {
-  return (dispatch) => {
-    dispatch(removeTask(id));
+  return async (dispatch) => {
+    const removedTask = await itemService.removeTask(id);
+    if (removedTask) {
+      dispatch(removeTask(id));
+    } else {
+      //TODO: Need to post the toast message
+    }
   };
 };
 
@@ -27,19 +50,29 @@ const removeAllTasks = () => ({
 });
 
 export const removeAllTasksAction = () => {
-  return (dispatch) => {
-    dispatch(removeAllTasks());
+  return async (dispatch) => {
+    const isRemoved = await itemService.removeAllTasks();
+    if (isRemoved) {
+      dispatch(removeAllTasks());
+    } else {
+      //TODO: Need to post the toast message
+    }
   };
 };
 
 const updateTask = (task) => ({
-  type: actionTypes.EDIT_TASK,
+  type: actionTypes.UPDATE_TASK,
   payload: task,
 });
 
 export const updateTaskAction = (task) => {
   return (dispatch) => {
-    dispatch(updateTask(task));
+    const updatedTask = itemService.updateTask(task);
+    if (updatedTask) {
+      dispatch(updateTask(task));
+    } else {
+      //TODO: Need to post the toast message
+    }
   };
 };
 
@@ -49,7 +82,23 @@ const setTasks = (tasks) => ({
 });
 
 export const setTasksAction = (tasks) => {
+  return async (dispatch) => {
+    const isUpdated = await itemService.updateTaskOrder(tasks);
+    if (isUpdated) {
+      dispatch(setTasks(tasks));
+    } else {
+      //TODO: Need to post the toast message
+    }
+  };
+};
+
+const setEditTask = (task) => ({
+  type: actionTypes.SET_EDIT_TASK,
+  payload: task,
+});
+
+export const setEditTaskAction = (task) => {
   return (dispatch) => {
-    dispatch(setTasks(tasks));
+    dispatch(setEditTask(task));
   };
 };
